+++ core/data/src/main/java/com/doyoonkim/data/di/RemoteModules.kt
@@ -1,8 +1,10 @@
 package com.doyoonkim.data.di
 
-import com.doyoonkim.data.repository.CampusRemoteRepository
-import com.doyoonkim.data.repository.ImageRepositoryImpl
-import com.doyoonkim.data.repository.RemoteRepositoryImpl
+import com.doyoonkim.data.repository.remote.CampusRemoteRepository
+import com.doyoonkim.data.repository.remote.ImageRepositoryImpl
+import com.doyoonkim.data.repository.remote.RemoteContentRepositoryImpl
+import com.doyoonkim.data.repository.remote.RemotePreferenceRepositoryImpl
+import com.doyoonkim.data.repository.remote.RemoteTokenRepositoryImpl
 import com.doyoonkim.domain.interfaces.CarrelStatusRemoteRepository
 import com.doyoonkim.domain.interfaces.ImageRemoteRepository
 import com.doyoonkim.domain.interfaces.NoticeRemoteRepository
@@ -17,7 +19,7 @@ import dagger.Module
 abstract class NoticeRemoteModule {
     @Binds
     abstract fun bindsNoticeRemoteRepo(
-        impl: RemoteRepositoryImpl
+        impl: RemoteContentRepositoryImpl
     ) : NoticeRemoteRepository
 }
 
@@ -25,7 +27,7 @@ abstract class NoticeRemoteModule {
 abstract class TipRemoteModule {
     @Binds
     abstract fun bindsTipRemoteRepo(
-        impl: RemoteRepositoryImpl
+        impl: RemoteContentRepositoryImpl
     ) : TipRemoteRepository
 }
 
@@ -33,7 +35,7 @@ abstract class TipRemoteModule {
 abstract class TokenRemoteModule {
     @Binds
     abstract fun bindsTokenRemoteRepo(
-        impl: RemoteRepositoryImpl
+        impl: RemoteTokenRepositoryImpl
     ) : TokenRemoteRepository
 }
 
@@ -41,12 +43,12 @@ abstract class TokenRemoteModule {
 abstract class PreferencesRemoteModule {
     @Binds
     abstract fun bindsTopicSubscriptionRemoteRepo(
-        impl: RemoteRepositoryImpl
+        impl: RemotePreferenceRepositoryImpl
     ) : TopicSubscriptionRemoteRepository
 
     @Binds
     abstract fun bindsUserReportRemoteRepo(
-        impl: RemoteRepositoryImpl
+        impl: RemotePreferenceRepositoryImpl
     ) : UserReportRemoteRepository
 }
