+++ core/data/src/main/java/com/doyoonkim/data/repository/local/LocalRepositoryImpl.kt
@@ -1,4 +1,4 @@
-package com.doyoonkim.data.repository
+package com.doyoonkim.data.repository.local
 
 import android.util.Log
 import androidx.room.withTransaction
