+++ core/data/src/main/java/com/doyoonkim/data/repository/remote/ImageRepositoryImpl.kt
@@ -1,4 +1,4 @@
-package com.doyoonkim.data.repository
+package com.doyoonkim.data.repository.remote
 
 import android.util.Log
 import com.doyoonkim.domain.interfaces.ImageRemoteRepository
